+++ src/configuration/settings/Settings.h
@@ -110,6 +110,16 @@ enum AdventureFieldDoorOverride
     AdventureFieldDoorOverrideAllClosed = 2,
 };
 
+
+enum MapButton
+{
+    MapButtonY = 0,
+    MapButtonSelect = 1,
+    MapButtonRightBumper = 2,
+    MapButtonLeftBumper = 3,
+    MapButtonDisabled = 3,
+};
+
 static std::string LeftTrim(std::string s)
 {
     s.erase(s.begin(), std::find_if(s.begin(), s.end(), [](unsigned char ch)
@@ -222,6 +232,8 @@ public:
     bool capsuleIndicator = true;
     bool fishIndicator = true;
     bool progressionIndicator = true;
+
+    MapButton mapButton = MapButtonY;
 
     NJS_COLOR arrowColor[6] = {
         {0xFF00FF00},